fix print array

+++ sort.h
@@ -30,7 +30,7 @@ void bubble_sort(int *array, size_t size);
 void insertion_sort_list(listint_t **list);
 void selection_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
-void quicksort(int *array, size_t low, size_t high);
-size_t lomuto(int *array, size_t low, size_t high);
+void quicksort(int *array, int low, int high);
+int lomuto(int *array, int low, int high);
 
 #endif
